Reusing search box more
