--- app/components/admin/search_box.html.erb
@@ -1,14 +1,0 @@
-<div class="admin-search-box">
-  <%= form_with method: :get, class: "admin-search-box__form" do |f| %>
-    <div class="admin-search-box__input">
-      <%= f.govuk_text_field(
-            param_name,
-            value: query,
-            label: { text: title, size: "s" },
-            hint: { text: hint },
-          )
-      %>
-    </div>
-    <%= f.govuk_submit "Search", classes: "admin-search-box__button", "data-test": "search-button" %>
-  <% end %>
-</div>
