--- app/components/admin/search_box.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module Admin
-  class SearchBox < BaseComponent
-    def initialize(query:, title: "Search", hint: nil, param_name: :query)
-      @query = query
-      @title = title
-      @hint = hint
-      @param_name = param_name
-    end
-
-  private
-
-    attr_reader :query, :title, :hint, :param_name
-  end
-end
