+++ app/views/admin/schools/index.html.erb
@@ -1,5 +1,5 @@
 <h1 class="govuk-heading-l">Schools</h1>
-<%= render Admin::SearchBox.new(query: @query, title: "Search schools", hint: "Enter the school’s name, URN or tutor email") %>
+<%= render SearchBox.new(query: @query, title: "Search schools", hint: "Enter the school’s name, URN or tutor email") %>
 
 
 <div class="nhsuk-table-container">
