+++ app/webpacker/styles/application.scss
@@ -25,7 +25,7 @@ $govuk-image-url-function: frontend-image-url;
 @import "subnav";
 @import "primary-nav";
 @import "school_search";
-@import "admin/search_box";
+@import "search_box";
 @import "aside";
 @import "vendor/nhs_tables";
