+++ app/webpacker/styles/search_box.scss
@@ -1,4 +1,4 @@
-.admin-search-box {
+.search-box {
   background-color: govuk-colour("light-grey");
   padding: govuk-spacing(4) govuk-spacing(3);
